// projects/b2b-tools/src/lib/components/selects/advanced-select/types/advanced-select-i18n.type.ts
export type AdvancedSelectLang = 'EN' | 'ES';

export interface AdvancedSelectI18n {
  modalTitle: string;
  placeholder: string;
  removeOption: string;
  moreOptions: string;
  more: string;
  clearSelection: string;
  advancedSearch: string;
  searchPlaceholder: string;
  noResults: string;
  advancedSearchPlaceholder: string;
  advancedSearchAction: string;
  select: string;
  idCode: string;
  descriptionOption: string;
  noModalResults: string;
  selected: string;
  cancel: string;
  confirmSelection: string;
}